New JHU-to-S3 ETL for IL only
