+++ covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17001.json_global
@@ -1,0 +1,1 @@
+{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":40,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":48,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":519,"deaths":5,"recovered":"<5"},"2020-09-28":{"confirmed":1149,"deaths":10,"recovered":"<5"},"2020-11-17":{"confirmed":3888,"deaths":38,"recovered":"<5"},"2021-01-06":{"confirmed":6870,"deaths":75,"recovered":"<5"},"2021-02-25":{"confirmed":7994,"deaths":114,"recovered":"<5"},"2021-04-16":{"confirmed":8178,"deaths":117,"recovered":"<5"},"2021-06-05":{"confirmed":8622,"deaths":125,"recovered":"<5"},"2021-07-25":{"confirmed":9386,"deaths":128,"recovered":"<5"},"2021-09-13":{"confirmed":12392,"deaths":145,"recovered":"<5"},"2021-11-02":{"confirmed":13401,"deaths":156,"recovered":"<5"},"2021-12-22":{"confirmed":15685,"deaths":164,"recovered":"<5"}}
