--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17003.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":20,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":37,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":83,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":220,"deaths":"<5","recovered":"<5"},"2021-01-06":{"confirmed":357,"deaths":7,"recovered":"<5"},"2021-02-25":{"confirmed":447,"deaths":8,"recovered":"<5"},"2021-04-16":{"confirmed":462,"deaths":10,"recovered":"<5"},"2021-06-05":{"confirmed":471,"deaths":11,"recovered":"<5"},"2021-07-25":{"confirmed":485,"deaths":11,"recovered":"<5"},"2021-09-13":{"confirmed":650,"deaths":11,"recovered":"<5"},"2021-11-02":{"confirmed":758,"deaths":12,"recovered":"<5"},"2021-12-22":{"confirmed":858,"deaths":13,"recovered":"<5"}}
