--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17005.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":6,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":12,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":60,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":330,"deaths":5,"recovered":"<5"},"2020-11-17":{"confirmed":842,"deaths":10,"recovered":"<5"},"2021-01-06":{"confirmed":1609,"deaths":16,"recovered":"<5"},"2021-02-25":{"confirmed":1902,"deaths":24,"recovered":"<5"},"2021-04-16":{"confirmed":1968,"deaths":24,"recovered":"<5"},"2021-06-05":{"confirmed":2085,"deaths":24,"recovered":"<5"},"2021-07-25":{"confirmed":2115,"deaths":24,"recovered":"<5"},"2021-09-13":{"confirmed":2399,"deaths":24,"recovered":"<5"},"2021-11-02":{"confirmed":2604,"deaths":26,"recovered":"<5"},"2021-12-22":{"confirmed":2957,"deaths":28,"recovered":"<5"}}
