--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17007.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":113,"deaths":9,"recovered":"<5"},"2020-06-20":{"confirmed":554,"deaths":20,"recovered":"<5"},"2020-08-09":{"confirmed":765,"deaths":23,"recovered":"<5"},"2020-09-28":{"confirmed":1128,"deaths":23,"recovered":"<5"},"2020-11-17":{"confirmed":3232,"deaths":29,"recovered":"<5"},"2021-01-06":{"confirmed":5239,"deaths":62,"recovered":"<5"},"2021-02-25":{"confirmed":5912,"deaths":71,"recovered":"<5"},"2021-04-16":{"confirmed":6348,"deaths":71,"recovered":"<5"},"2021-06-05":{"confirmed":6813,"deaths":75,"recovered":"<5"},"2021-07-25":{"confirmed":6911,"deaths":80,"recovered":"<5"},"2021-09-13":{"confirmed":7502,"deaths":81,"recovered":"<5"},"2021-11-02":{"confirmed":8064,"deaths":84,"recovered":"<5"},"2021-12-22":{"confirmed":9697,"deaths":89,"recovered":"<5"}}
