--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17009.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":10,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":14,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":62,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":236,"deaths":"<5","recovered":"<5"},"2021-01-06":{"confirmed":475,"deaths":5,"recovered":"<5"},"2021-02-25":{"confirmed":660,"deaths":6,"recovered":"<5"},"2021-04-16":{"confirmed":684,"deaths":6,"recovered":"<5"},"2021-06-05":{"confirmed":725,"deaths":6,"recovered":"<5"},"2021-07-25":{"confirmed":800,"deaths":6,"recovered":"<5"},"2021-09-13":{"confirmed":1009,"deaths":7,"recovered":"<5"},"2021-11-02":{"confirmed":1076,"deaths":11,"recovered":"<5"},"2021-12-22":{"confirmed":1194,"deaths":13,"recovered":"<5"}}
