+++ covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17031.json
@@ -1,1 +1,1 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":27,"deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":38668,"deaths":1673,"recovered":"<5"},"2020-06-20":{"confirmed":86885,"deaths":4390,"recovered":"<5"},"2020-08-09":{"confirmed":111447,"deaths":4928,"recovered":"<5"},"2020-09-28":{"confirmed":144315,"deaths":5217,"recovered":"<5"},"2020-11-17":{"confirmed":260004,"deaths":5855,"recovered":"<5"},"2021-01-06":{"confirmed":406674,"deaths":8101,"recovered":"<5"},"2021-02-25":{"confirmed":472206,"deaths":9317,"recovered":"<5"},"2021-04-16":{"confirmed":519374,"deaths":9807,"recovered":"<5"},"2021-06-05":{"confirmed":554813,"deaths":10357,"recovered":"<5"},"2021-07-25":{"confirmed":561168,"deaths":10553,"recovered":"<5"},"2021-09-13":{"confirmed":605061,"deaths":10877,"recovered":"<5"},"2021-11-02":{"confirmed":644255,"deaths":11303,"recovered":"<5"},"2021-12-22":{"confirmed":749730,"deaths":11811,"recovered":"<5"}}
+{"2020-01-22":{"confirmed":"<5","deaths":"<5","C":"<5","D":"<5"},"2020-03-12":{"confirmed":27,"deaths":"<5","C":27,"D":"<5"},"2020-05-01":{"confirmed":38668,"deaths":1673,"C":38668,"D":1673},"2020-06-20":{"confirmed":86885,"deaths":4390,"C":86885,"D":4390},"2020-08-09":{"confirmed":111447,"deaths":4928,"C":111447,"D":4928},"2020-09-28":{"confirmed":144315,"deaths":5217,"C":144315,"D":5217},"2020-11-17":{"confirmed":260004,"deaths":5855,"C":260004,"D":5855},"2021-01-06":{"confirmed":406674,"deaths":8101,"C":406674,"D":8101},"2021-02-25":{"confirmed":472206,"deaths":9317,"C":472206,"D":9317},"2021-04-16":{"confirmed":519374,"deaths":9807,"C":519374,"D":9807},"2021-06-05":{"confirmed":554813,"deaths":10357,"C":554813,"D":10357},"2021-07-25":{"confirmed":561168,"deaths":10553,"C":561168,"D":10553},"2021-09-13":{"confirmed":605061,"deaths":10877,"C":605061,"D":10877},"2021-11-02":{"confirmed":644255,"deaths":11303,"C":644255,"D":11303},"2021-12-22":{"confirmed":749730,"deaths":11811,"C":749730,"D":11811}}
