+++ covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17031.json_global
@@ -1,0 +1,1 @@
+{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":27,"deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":38668,"deaths":1673,"recovered":"<5"},"2020-06-20":{"confirmed":86885,"deaths":4390,"recovered":"<5"},"2020-08-09":{"confirmed":111447,"deaths":4928,"recovered":"<5"},"2020-09-28":{"confirmed":144315,"deaths":5217,"recovered":"<5"},"2020-11-17":{"confirmed":260004,"deaths":5855,"recovered":"<5"},"2021-01-06":{"confirmed":406674,"deaths":8101,"recovered":"<5"},"2021-02-25":{"confirmed":472206,"deaths":9317,"recovered":"<5"},"2021-04-16":{"confirmed":519374,"deaths":9807,"recovered":"<5"},"2021-06-05":{"confirmed":554813,"deaths":10357,"recovered":"<5"},"2021-07-25":{"confirmed":561168,"deaths":10553,"recovered":"<5"},"2021-09-13":{"confirmed":605061,"deaths":10877,"recovered":"<5"},"2021-11-02":{"confirmed":644255,"deaths":11303,"recovered":"<5"},"2021-12-22":{"confirmed":749730,"deaths":11811,"recovered":"<5"}}
