--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17033.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":10,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":17,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":29,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":458,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":897,"deaths":8,"recovered":"<5"},"2021-01-06":{"confirmed":1643,"deaths":19,"recovered":"<5"},"2021-02-25":{"confirmed":2003,"deaths":21,"recovered":"<5"},"2021-04-16":{"confirmed":2090,"deaths":25,"recovered":"<5"},"2021-06-05":{"confirmed":2142,"deaths":26,"recovered":"<5"},"2021-07-25":{"confirmed":2168,"deaths":26,"recovered":"<5"},"2021-09-13":{"confirmed":2891,"deaths":30,"recovered":"<5"},"2021-11-02":{"confirmed":3379,"deaths":30,"recovered":"<5"},"2021-12-22":{"confirmed":3790,"deaths":32,"recovered":"<5"}}
