--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17035.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":10,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":55,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":218,"deaths":5,"recovered":"<5"},"2020-11-17":{"confirmed":496,"deaths":6,"recovered":"<5"},"2021-01-06":{"confirmed":996,"deaths":14,"recovered":"<5"},"2021-02-25":{"confirmed":1186,"deaths":18,"recovered":"<5"},"2021-04-16":{"confirmed":1279,"deaths":19,"recovered":"<5"},"2021-06-05":{"confirmed":1298,"deaths":19,"recovered":"<5"},"2021-07-25":{"confirmed":1309,"deaths":19,"recovered":"<5"},"2021-09-13":{"confirmed":1693,"deaths":23,"recovered":"<5"},"2021-11-02":{"confirmed":1933,"deaths":27,"recovered":"<5"},"2021-12-22":{"confirmed":2274,"deaths":29,"recovered":"<5"}}
