--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17037.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":105,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":517,"deaths":18,"recovered":"<5"},"2020-08-09":{"confirmed":936,"deaths":30,"recovered":"<5"},"2020-09-28":{"confirmed":1590,"deaths":38,"recovered":"<5"},"2020-11-17":{"confirmed":4146,"deaths":44,"recovered":"<5"},"2021-01-06":{"confirmed":6905,"deaths":81,"recovered":"<5"},"2021-02-25":{"confirmed":8226,"deaths":112,"recovered":"<5"},"2021-04-16":{"confirmed":9216,"deaths":118,"recovered":"<5"},"2021-06-05":{"confirmed":10081,"deaths":120,"recovered":"<5"},"2021-07-25":{"confirmed":10235,"deaths":122,"recovered":"<5"},"2021-09-13":{"confirmed":11399,"deaths":126,"recovered":"<5"},"2021-11-02":{"confirmed":12390,"deaths":134,"recovered":"<5"},"2021-12-22":{"confirmed":14946,"deaths":154,"recovered":"<5"}}
