--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17039.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":33,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":125,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":510,"deaths":8,"recovered":"<5"},"2021-01-06":{"confirmed":997,"deaths":18,"recovered":"<5"},"2021-02-25":{"confirmed":1277,"deaths":22,"recovered":"<5"},"2021-04-16":{"confirmed":1379,"deaths":23,"recovered":"<5"},"2021-06-05":{"confirmed":1569,"deaths":29,"recovered":"<5"},"2021-07-25":{"confirmed":1606,"deaths":29,"recovered":"<5"},"2021-09-13":{"confirmed":1889,"deaths":30,"recovered":"<5"},"2021-11-02":{"confirmed":2121,"deaths":30,"recovered":"<5"},"2021-12-22":{"confirmed":2803,"deaths":32,"recovered":"<5"}}
