--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17101.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":5,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":47,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":259,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":634,"deaths":8,"recovered":"<5"},"2021-01-06":{"confirmed":2029,"deaths":20,"recovered":"<5"},"2021-02-25":{"confirmed":2288,"deaths":24,"recovered":"<5"},"2021-04-16":{"confirmed":2389,"deaths":25,"recovered":"<5"},"2021-06-05":{"confirmed":2410,"deaths":25,"recovered":"<5"},"2021-07-25":{"confirmed":2436,"deaths":27,"recovered":"<5"},"2021-09-13":{"confirmed":2882,"deaths":30,"recovered":"<5"},"2021-11-02":{"confirmed":3391,"deaths":35,"recovered":"<5"},"2021-12-22":{"confirmed":3788,"deaths":37,"recovered":"<5"}}
