--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17103.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":23,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":94,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":171,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":383,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":1520,"deaths":5,"recovered":"<5"},"2021-01-06":{"confirmed":2821,"deaths":32,"recovered":"<5"},"2021-02-25":{"confirmed":3342,"deaths":43,"recovered":"<5"},"2021-04-16":{"confirmed":3940,"deaths":50,"recovered":"<5"},"2021-06-05":{"confirmed":4192,"deaths":53,"recovered":"<5"},"2021-07-25":{"confirmed":4225,"deaths":54,"recovered":"<5"},"2021-09-13":{"confirmed":4496,"deaths":56,"recovered":"<5"},"2021-11-02":{"confirmed":4872,"deaths":62,"recovered":"<5"},"2021-12-22":{"confirmed":5727,"deaths":67,"recovered":"<5"}}
