--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17105.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":20,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":37,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":115,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":403,"deaths":6,"recovered":"<5"},"2020-11-17":{"confirmed":1725,"deaths":15,"recovered":"<5"},"2021-01-06":{"confirmed":3246,"deaths":53,"recovered":"<5"},"2021-02-25":{"confirmed":4164,"deaths":75,"recovered":"<5"},"2021-04-16":{"confirmed":4580,"deaths":83,"recovered":"<5"},"2021-06-05":{"confirmed":4888,"deaths":89,"recovered":"<5"},"2021-07-25":{"confirmed":4931,"deaths":92,"recovered":"<5"},"2021-09-13":{"confirmed":5559,"deaths":94,"recovered":"<5"},"2021-11-02":{"confirmed":6050,"deaths":99,"recovered":"<5"},"2021-12-22":{"confirmed":7053,"deaths":113,"recovered":"<5"}}
