--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17107.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":7,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":13,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":112,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":468,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":1239,"deaths":7,"recovered":"<5"},"2021-01-06":{"confirmed":2842,"deaths":41,"recovered":"<5"},"2021-02-25":{"confirmed":3304,"deaths":53,"recovered":"<5"},"2021-04-16":{"confirmed":3628,"deaths":56,"recovered":"<5"},"2021-06-05":{"confirmed":3959,"deaths":62,"recovered":"<5"},"2021-07-25":{"confirmed":4046,"deaths":65,"recovered":"<5"},"2021-09-13":{"confirmed":4692,"deaths":73,"recovered":"<5"},"2021-11-02":{"confirmed":4994,"deaths":76,"recovered":"<5"},"2021-12-22":{"confirmed":5908,"deaths":77,"recovered":"<5"}}
