--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17109.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":18,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":101,"deaths":15,"recovered":"<5"},"2020-08-09":{"confirmed":143,"deaths":15,"recovered":"<5"},"2020-09-28":{"confirmed":482,"deaths":15,"recovered":"<5"},"2020-11-17":{"confirmed":1337,"deaths":26,"recovered":"<5"},"2021-01-06":{"confirmed":2028,"deaths":36,"recovered":"<5"},"2021-02-25":{"confirmed":2314,"deaths":40,"recovered":"<5"},"2021-04-16":{"confirmed":2561,"deaths":43,"recovered":"<5"},"2021-06-05":{"confirmed":2897,"deaths":51,"recovered":"<5"},"2021-07-25":{"confirmed":3026,"deaths":51,"recovered":"<5"},"2021-09-13":{"confirmed":3497,"deaths":54,"recovered":"<5"},"2021-11-02":{"confirmed":3801,"deaths":64,"recovered":"<5"},"2021-12-22":{"confirmed":4507,"deaths":70,"recovered":"<5"}}
