--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17141.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":122,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":234,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":410,"deaths":5,"recovered":"<5"},"2020-09-28":{"confirmed":728,"deaths":6,"recovered":"<5"},"2020-11-17":{"confirmed":2326,"deaths":16,"recovered":"<5"},"2021-01-06":{"confirmed":4306,"deaths":59,"recovered":"<5"},"2021-02-25":{"confirmed":4987,"deaths":73,"recovered":"<5"},"2021-04-16":{"confirmed":5739,"deaths":78,"recovered":"<5"},"2021-06-05":{"confirmed":6190,"deaths":83,"recovered":"<5"},"2021-07-25":{"confirmed":6275,"deaths":84,"recovered":"<5"},"2021-09-13":{"confirmed":6953,"deaths":85,"recovered":"<5"},"2021-11-02":{"confirmed":7726,"deaths":92,"recovered":"<5"},"2021-12-22":{"confirmed":9675,"deaths":106,"recovered":"<5"}}
