--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17143.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":93,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":423,"deaths":23,"recovered":"<5"},"2020-08-09":{"confirmed":1607,"deaths":35,"recovered":"<5"},"2020-09-28":{"confirmed":3510,"deaths":50,"recovered":"<5"},"2020-11-17":{"confirmed":7281,"deaths":109,"recovered":"<5"},"2021-01-06":{"confirmed":13745,"deaths":202,"recovered":"<5"},"2021-02-25":{"confirmed":16848,"deaths":260,"recovered":"<5"},"2021-04-16":{"confirmed":20980,"deaths":277,"recovered":"<5"},"2021-06-05":{"confirmed":23418,"deaths":334,"recovered":"<5"},"2021-07-25":{"confirmed":23620,"deaths":346,"recovered":"<5"},"2021-09-13":{"confirmed":25476,"deaths":358,"recovered":"<5"},"2021-11-02":{"confirmed":27246,"deaths":374,"recovered":"<5"},"2021-12-22":{"confirmed":31889,"deaths":401,"recovered":"<5"}}
