--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17145.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":24,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":44,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":184,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":353,"deaths":15,"recovered":"<5"},"2020-11-17":{"confirmed":793,"deaths":16,"recovered":"<5"},"2021-01-06":{"confirmed":2388,"deaths":50,"recovered":"<5"},"2021-02-25":{"confirmed":3033,"deaths":57,"recovered":"<5"},"2021-04-16":{"confirmed":3130,"deaths":59,"recovered":"<5"},"2021-06-05":{"confirmed":3194,"deaths":60,"recovered":"<5"},"2021-07-25":{"confirmed":3266,"deaths":61,"recovered":"<5"},"2021-09-13":{"confirmed":4046,"deaths":69,"recovered":"<5"},"2021-11-02":{"confirmed":4379,"deaths":76,"recovered":"<5"},"2021-12-22":{"confirmed":4751,"deaths":80,"recovered":"<5"}}
