--- covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17147.json
@@ -1,1 +1,0 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":7,"deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":16,"deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":61,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":139,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":621,"deaths":5,"recovered":"<5"},"2021-01-06":{"confirmed":1090,"deaths":8,"recovered":"<5"},"2021-02-25":{"confirmed":1318,"deaths":14,"recovered":"<5"},"2021-04-16":{"confirmed":1415,"deaths":14,"recovered":"<5"},"2021-06-05":{"confirmed":1520,"deaths":14,"recovered":"<5"},"2021-07-25":{"confirmed":1549,"deaths":14,"recovered":"<5"},"2021-09-13":{"confirmed":1868,"deaths":14,"recovered":"<5"},"2021-11-02":{"confirmed":2055,"deaths":14,"recovered":"<5"},"2021-12-22":{"confirmed":2764,"deaths":16,"recovered":"<5"}}
