+++ covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17149.json
@@ -1,1 +1,1 @@
-{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":21,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":134,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":860,"deaths":15,"recovered":"<5"},"2021-01-06":{"confirmed":1452,"deaths":39,"recovered":"<5"},"2021-02-25":{"confirmed":1687,"deaths":50,"recovered":"<5"},"2021-04-16":{"confirmed":1760,"deaths":51,"recovered":"<5"},"2021-06-05":{"confirmed":1804,"deaths":53,"recovered":"<5"},"2021-07-25":{"confirmed":1971,"deaths":53,"recovered":"<5"},"2021-09-13":{"confirmed":2476,"deaths":56,"recovered":"<5"},"2021-11-02":{"confirmed":2683,"deaths":58,"recovered":"<5"},"2021-12-22":{"confirmed":3405,"deaths":58,"recovered":"<5"}}
+{"2020-01-22":{"confirmed":"<5","deaths":"<5","C":"<5","D":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","C":"<5","D":"<5"},"2020-05-01":{"confirmed":"<5","deaths":"<5","C":"<5","D":"<5"},"2020-06-20":{"confirmed":"<5","deaths":"<5","C":"<5","D":"<5"},"2020-08-09":{"confirmed":21,"deaths":"<5","C":21,"D":"<5"},"2020-09-28":{"confirmed":134,"deaths":"<5","C":134,"D":"<5"},"2020-11-17":{"confirmed":860,"deaths":15,"C":860,"D":15},"2021-01-06":{"confirmed":1452,"deaths":39,"C":1452,"D":39},"2021-02-25":{"confirmed":1687,"deaths":50,"C":1687,"D":50},"2021-04-16":{"confirmed":1760,"deaths":51,"C":1760,"D":51},"2021-06-05":{"confirmed":1804,"deaths":53,"C":1804,"D":53},"2021-07-25":{"confirmed":1971,"deaths":53,"C":1971,"D":53},"2021-09-13":{"confirmed":2476,"deaths":56,"C":2476,"D":56},"2021-11-02":{"confirmed":2683,"deaths":58,"C":2683,"D":58},"2021-12-22":{"confirmed":3405,"deaths":58,"C":3405,"D":58}}
