+++ covid19-etl/tests/test_jhu_to_s3_output/time_series/county/17149.json_global
@@ -1,0 +1,1 @@
+{"2020-01-22":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-03-12":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-05-01":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-06-20":{"confirmed":"<5","deaths":"<5","recovered":"<5"},"2020-08-09":{"confirmed":21,"deaths":"<5","recovered":"<5"},"2020-09-28":{"confirmed":134,"deaths":"<5","recovered":"<5"},"2020-11-17":{"confirmed":860,"deaths":15,"recovered":"<5"},"2021-01-06":{"confirmed":1452,"deaths":39,"recovered":"<5"},"2021-02-25":{"confirmed":1687,"deaths":50,"recovered":"<5"},"2021-04-16":{"confirmed":1760,"deaths":51,"recovered":"<5"},"2021-06-05":{"confirmed":1804,"deaths":53,"recovered":"<5"},"2021-07-25":{"confirmed":1971,"deaths":53,"recovered":"<5"},"2021-09-13":{"confirmed":2476,"deaths":56,"recovered":"<5"},"2021-11-02":{"confirmed":2683,"deaths":58,"recovered":"<5"},"2021-12-22":{"confirmed":3405,"deaths":58,"recovered":"<5"}}
